requirements.txt
Django==4.2.7
Pillow>=10.2.0
reportlab==4.0.7
qrcode==7.4.2
django-crispy-forms==2.1
crispy-bootstrap5==0.7
python-dateutil==2.8.2
PyMySQL==1.1.0
# mysqlclient==2.2.0  # Nécessite des outils de compilation C++ sur Windows

